@karinorman I've added a partial debug file here
Just took a quick dig, the EcoSimR code is pretty crude.  So far, looks like weird things are going on with sim9, sim1 (and I think other methods) might not have this issue.

sim9 is pretty opaque, but the issue has to do with sim9_single, where the random sampling should be occurring (via calls to R's "sample" function).  Somehow that function behaves differently on your dummy data than it does on the dummy data that function provides (in which the simulation is different in repeated calls).

Anyway, take a look over this script, hopefully it at least makes some sense to you what I'm doing / testing for.  we can dig more tomorrow.